+++ src/Command/ApiGatewaysCommand.php
@@ -18,7 +18,7 @@ class ApiGatewaysCommand extends Command
 {
     public function __construct(
         private readonly ApiClient $apiClient,
-        private readonly array $applicationStatus,
+        private readonly array $statuses,
     ) {
         parent::__construct();
     }
@@ -26,14 +26,16 @@ class ApiGatewaysCommand extends Command
     protected function configure(): void
     {
         $this
-            ->addOption('filterStatus', null, InputOption::VALUE_NONE, 'Filter based on configured statuses: '.implode(',', $this->applicationStatus))
+            ->addOption('filter-status', null, InputOption::VALUE_NONE, 'Filter based on configured statuses: '.implode(',', $this->statuses))
+            ->addOption('id', null, InputOption::VALUE_REQUIRED, 'Show details for gateway with this ID', -1)
         ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $io = new SymfonyStyle($input, $output);
-        $filter = $input->getOption('filterStatus');
+        $filter = $input->getOption('filter-status');
+        $id = (int) $input->getOption('id');
 
         $gateways = $this->apiClient->getGateways($filter);
 
@@ -42,9 +44,24 @@ class ApiGatewaysCommand extends Command
             $io->writeln($msg);
         }
 
+        // Show details for one of the gateways.
+        if ($id >= 0) {
+            $found = false;
+            foreach ($gateways as $gateway) {
+                if ($gateway->id === $id) {
+                    $output->writeln('');
+                    $output->writeln(json_encode($gateway, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE));
+                    $found = true;
+                }
+            }
+            if (!$found) {
+                $io->error('Gateway with the id '.$id.' not found');
+            }
+        }
+
         $msg = count($gateways);
         if ($filter) {
-            $msg .= sprintf(' gateways found (filter on status "%s")', implode(',', $this->applicationStatus));
+            $msg .= sprintf(' gateways found (filter on status "%s")', implode(',', $this->statuses));
         } else {
             $msg .= ' applications found';
         }